feat: animateBlocksTransition 훅, isInBound util 구현
