+++ frontend/src/dndTimeline/types/blockType.type.ts
@@ -1,0 +1,8 @@
+import type { Position } from './position.type';
+import type { Size } from '@/types/workCard.type';
+
+export interface BlockType {
+  id: string | number;
+  position: Position;
+  size: Size;
+}
